Split Gen* tests to have core GenShader + Gen<language> tests.
