+++ source/MaterialXTest/Render.cpp
